refactor: moved files from db folder to constants
