+++ src/constants/defaultOhlcPairs.ts
