+++ src/constants/index.ts
@@ -1,3 +1,8 @@
+export { BANKS_DATA } from './banksData';
+export { DEFAULT_CHART_DATA } from './defaultChartData';
+export { DEFAULT_CURRENCIES } from './defaultCurrencies';
+export { DEFAULT_OHLC_PAIRS } from './defaultOhlcPairs';
+
 export const CURRENCY_MODAL_QUERY_KEY = 'currency';
 
 export const {
@@ -5,6 +10,6 @@ export const {
   CURRENCY_API_BASE,
   COIN_API_KEY,
   COIN_API_BASE,
-  AUTOR_GITHUB,
+  AUTHOR_GITHUB,
   MODSEN_LINK,
 } = process.env;
