+++ src/pages/Timeline/Timeline.tsx
@@ -4,7 +4,7 @@ import CandleStick from '@/components/CandleStick';
 import HistoryForm from '@/components/HistoryForm';
 import { Button, Modal } from '@/components/UI';
 import UpdateDate from '@/components/UpdateDate';
-import { DEFAULT_OHLC_PAIRS } from '@/db/defaultOhlcPairs';
+import { DEFAULT_OHLC_PAIRS } from '@/constants';
 import { CandleStickData, OhlcvResponseType } from '@/types';
 import { getTimeFromDate } from '@/utils';
 import { ohlcvResponseToChartData } from '@/utils/ohlcvResponseToChartData';
