+++ src/utils/filterBanksByCurrencies.ts
@@ -1,4 +1,4 @@
-import { BANKS_DATA } from '@/db/banksData';
+import { BANKS_DATA } from '@/constants';
 
 export function filterBanksByCurrencies(target: string, banks = BANKS_DATA) {
   return banks.filter(({ currencies }) => {
