+++ src/utils/getBanksCurrencies.ts
@@ -1,4 +1,4 @@
-import { BANKS_DATA } from '@/db/banksData';
+import { BANKS_DATA } from '@/constants';
 
 export function getBanksCurrencies(banksData: typeof BANKS_DATA) {
   const currenciesArr = banksData.map(({ currencies }) => currencies).flat();
